+++ lib/resync/client/mixins/resource_client_delegate.rb
@@ -1,9 +1,10 @@
+require 'resync'
 require_relative 'client_delegator'
 
-# A resource container that is capable of providing those resources with a {Client}
 module Resync
   class Client
     module Mixins
+      # A resource container that is capable of providing those resources with a {Client}
       module ResourceClientDelegate
         prepend ClientDelegator
 
@@ -15,5 +16,9 @@ module Resync
         end
       end
     end
+  end
+
+  class BaseResourceList
+    prepend Client::Mixins::ResourceClientDelegate
   end
 end