+++ lib/resync/client/mixins/zipped_resource.rb
@@ -1,10 +1,10 @@
 require_relative '../zip'
 require_relative 'downloadable'
 
-# A resource that refers to a zipped bitstream package.
 module Resync
   class Client
     module Mixins
+      # A resource that refers to a zipped bitstream package.
       module ZippedResource
         prepend Downloadable
 
@@ -12,6 +12,7 @@ module Resync
         # it to a temporary file if necessary.
         # @return [Resync::Client::Zip::ZipPackage] the zipped contents of this resource
         def zip_package
+          puts self
           @zip_package ||= Resync::Client::Zip::ZipPackage.new(download_to_temp_file)
         end
       end